Get it closer to building
Mostly updating OpenGL pieces that used to be extensions and are now
part of Core.

This is finicky, because it really comes down to a question of driver
versions. It sounds like that was a big part of the original issue that
drove the project into obscurity in the first place.