Start implementing event handlers
Have mouse and keyboard buttons.

This version doesn't build, because it's trying to link to a
native JNI wrapper from a different platform. Actually, there are almost
definitely lots of other problems (I just ganked out a couple of major
foundation pieces), but that's the one that's blocking me immediately.