Docstrings illegal in interface implementations
In defrecord. So moved a "Do I want to do this?" docstring that broke
the build.

Also jiggered the version of the LWJGL snapshot version I'm using,
because that's the version I wound up installing w/ mvn on this
VM. Silly reason, but that's the way it worked out for now.