Finish getting Asteroids to compile
It still can't be even close to a working demo. Desperately needs a real
key press event handler. But at least it builds.

Get pong to compile, w/ same limitations.

Clean up a warning about the name of a var that looks dynamic. I'm not
sure that declaring it dynamic was the way to go, but I want to get rid
of all dynamic vars anyway. So this seems like a reasonable choice.

Though it would have arguably been better to leave the warning as a flag
to help make sure I don't forget about it later.

Start trying to do something sensible w/ Tetris. Get back around to
keyboard events and handlers, and changes in underlying architecture
between lwjgl versions.

This is a horrible place to stop, but I think this is all I'm going to
manage on this flight. Especially since we're starting our descent.