Switch to a Manager for tracking Apps.
I'm not happy with the nomenclature. But I think the basic idea is
valid.

The System should own a StageManager (or something along those lines),
which manages the individual Apps (which each include their own Windows,
along with all the other pieces that are involved).

And then individual orangelets (I'm still working on that name) can
register with an App as the game du jour.

At least that's my current plan.