Fatal IO Error 11
This was happening after calling (reset) 2 or 3 times. Couldn't get a
core dump (apparently that's one of the 'typical' symptoms), and google
failed me pretty much completely.

There was a java error log that pointed back to glfwTerminate. So took
that out of the equation. Just call initialize once, then never call
terminate.

It feels dirty, but seems to be a valid work-around for now