Get it to build again
This was mostly mechanical, but it did involve digging into some of the
original functionality to get a grasp on what the basic event loop is
actually doing.

I'm not sure that this version's actually doing anything better than the
previous version (since I'm really just punting on any chance at
resuming from a pause), but...it builds.